Make Legend.create() aware of existing groups in Layer Tree
That means, if groups like (Tables, Domains) are present in the Layer Tree and the user loads more layers using Project Generator, new layers go to the corresponding group, i.e., the plugin won't duplicate groups in the Layer Tree.

Additionally, reversing the order of points and polygons in Legend.py and using group.insertLayer instead of group.addLayer in Legend.create() avoids getting new layers loaded at the bottom of the Layer Tree, which is happening when loading new layers using Project Generator into a project with existing layers and groups.