Display current external attributes file
In the lable updating when changing the attributes (toml) file. The attributes (toml) file is reloaded depending on the model.
It's stored as configuration for every model. If the model has no toml-file the lable is empty as well.
On valid ili-files it takes the last model for modelvalue (the one that would been taken as the model if no one has been entered).