Add the simple tests with ili2db4 created geopackage.
Fix comparison, since the simple road data contains more than one element with the same tag name. But the id can be different to the original XTF.

Renamed attribut with childs, since they are child elements an not attributes.

It does ignore the child element 'Geometry' as well in the end, since the nested values are numbers with decimal places or not. These places divert from the created XTF to the current one. So a simple top level string comparison would fail. (it would make sence to improve the xtf-comparison function in future to have a comparison of every nested child by it's value).