Fix modeldir support of % values
The old approach was locally searching for .ili files in the model directories
and recursively adding all of them. The setting actually supports a couple of
magic strings like `%XTF_DIR` and similar which got lost this way because
we couldn't glob them.

This approach always adds all directories from the setting and adds
additional subdirs only if they contain .ili files

Fixes #464